requirements.txt
Flask==3.0.0
flask-swagger-ui==5.21.0
python-dotenv==1.0.1
pytest==7.4.3
pytest-cov==4.1.0
testcontainers==4.10.0
SQLAlchemy==2.0.42
bcrypt==4.3.0